Exercise 2
I got behind on the videos, so this exercise is definitely late, but I
wanted to go through and watch the videos to catch up anyway, despite
it being far from the exercise 2 deadline. Although I was able to fix
one image, I couldn’t figure out the other two. I will come in to
office hours to address this next week. I apologize for the delay.